Add authorized routes and useUser
This blocks non-authorized people from accessing the Edit screen,
presumably (I can't test yet because of errors on D3Chart). Unauthorized
people shouldn't be allowed to Edit the tree because:
- We need to be able to track who added what content, so author expertise
could be tagged onto the content, and so people spamming troll content
could be identified and kicked, in case that happens.
- We can't allow just anyone to update or remove items on the tree
because this might cause loss of data.