Move ToolButtons to Edit
Just as I was about to implement the onClick handlers for the ToolButtons
I realized it didn't make sense to send those handlers to NodeDescription
when they only make sense at the Edit page. Sending the buttons as
children make more sense so that the handlers only need to be inside
the Edit page.