Implement the Submit functionality
This is multiple commits squashed into one. I'm keeping some of the
messages that might be important in the future. The Submit feature
send branches to the admin's page, where they can confirm if a branch
should be added to the main tree.

Add preliminary validation and alert

I plan to use RHF and a different Alert Dialog for this later on.

Add isRooted property to nodes

Previously, I added a "rootedNodes" array property to the branch
thinking that I will use this upon submission to determine which
parts of the branch are rooted to the universal tree and which are not.
But upon further thinking, doing this would require traversing the whole
branch and more complex calculations which will probably affect
performance. Having a boolean on each node however makes things simple:
all nodes that have isRooted === true would be part of the rooted branch
and the rest would be part of the hanging branch.

Implement Skill Node validation

For now, skill node validation merely consists of informing the user
about disconnected skill nodes which instead of being added to the
main tree display will be added to a yet-to-be-implemented Requests
page. Requests are a collection of skill nodes or branches that we
want to be connected to the main tree via modules that some more
knowledgeable user could make for us.

Default skill node validity to true

Without this we can't submit, because the only thing that sets the
validity to true is a button in the alert box that only appears if the
submissions are questionable.

Rename and remove some functions

The getDisconnectedNodes function reads like it was written before
the "is_rooted" variable existed, and it complicates the process of
identifying disconnected nodes. I'm not sure though so I merely commented
it out.

Update handleSubmit and its helpers

Previously, the handleSubmit wasn't working correctly with the invalid
skill nodes alert. Pressing No on the alert still resulted in submission
because of stale state. This fixes that. I also updated the module
validation since we should be checking for hanging modules, not unrooted
modules.